:fire: :truck: refactor(CartViewModel): CartViewModel 제거, ShoppingCartViewModel 로 이름 변경
